cleaned up a little code
