+++ src/App.js
@@ -12,11 +12,10 @@ import Navbar from './components/Navbar';
 import Greece from './components/pages/Destinations/Greece';
 import Brazil from './components/pages/Destinations/Brazil';
 import Mexico from './components/pages/Destinations/Mexico';
-import SearchBar from './components/SearchBar';
-import SearchResult from './Data.json';
+
 
 const App = () => {
-//const title = 'You Can Know!'
+// const title = 'You Can Know!'
 
   return (
     
@@ -35,8 +34,7 @@ const App = () => {
         <Route path='/greece' component={Greece}/>  
         <Route path='/brazil' component={Brazil}/>      
         <Route path='/mexico' component={Mexico}/>      
-        <div className='App'>
-        <SearchBar placeholder='Enter Search Text Here...' /></div>
+
         </Switch>
       </Router>
  );
